+++ akamai-staging.js
@@ -6,114 +6,141 @@ const debug = require('debug')('staging');
 // Platform independent end-of-line character
 const endOfLine = require('os').EOL;
 
-// Synchronous DNS resolver
+// akamai-staging's own DNS library
+const dnsLib = require('./dns-core');
+
+// node DNS library
 const dns = require('dns');
 
+// Initialise wildcard string parser
+const matcher = require('multimatch');
+
+// Array of Akamai edge domains
+const edgeDomains = [
+    '*.edgekey.net',
+    '*.edgesuite.net',
+    '*.akamaiedge.net'
+];
+
+
 function getStagingIPAddress(hostname){
-  debug('getStagingIPAddress() Entry');
-  // Resolve the hostname to its aliases
-  debug('Calling dns.resolveCname(%s)', hostname);
+    debug('getStagingIPAddress() Entry');
+    // Resolve the hostname to its aliases
+    debug('Calling resolveCname(%s)', hostname);
 
 
-  dns.resolveCname(hostname, function (err, aliases) {
-    if (err) {
-      debug('An error occurred in dns.resolveCname(%s): %O', hostname, err);
-      debug('No aliases returned. The record "%s" might not be a CNAME.', hostname);
-      console.log('# [%s] did not resolve to a CNAME record', hostname);
-    } else {
-      debug('Processing returned aliases: %O', aliases);
+    dnsLib.resolveCname(hostname, (response) => {
+        if (response.success === false) {
+            debug('An error [%s] occurred in resolveCname(%s): %O', response.message, hostname, response.error);
+            debug('No aliases returned. The record "%s" might not be a CNAME.', hostname);
+            console.error('# [%s] did not resolve to a CNAME record', hostname);
 
-      // Pick the 1st alias
-      debug('Using the 1st alias "%s"', aliases[0]);
-      var alias = aliases[0];
+        } else {
 
-      // Break down the alias hostname into its constituent parts/segments
-      var segments = alias.split('.');
-      debug('Split into segments: %O', segments);
+            // Check if any alias matches an Akamai edge domain
+            debug('Processing returned aliases: %O', response.aliases);
 
-      // append '-staging' to the penultimate segment
-      var stagingDomain = segments[segments.length - 2] + '-staging';
-      debug('Going to be using "%s" instead of "%s"', stagingDomain, segments[segments.length - 2]);
+            let matchingAliases = matcher(response.aliases, edgeDomains);
 
-      // Replace the penultimate segment with the '-staging' variant
-      segments[segments.length - 2] = stagingDomain;
-      debug('Edited segments: %O', segments);
+            if (matchingAliases.length > 0) {
+                // There's at least one alias that matches an Akamai edge domain pattern
+                debug('These look like Akamai domains: %O', matchingAliases);
 
-      // Variable where we're going to store the fully qualified Staging variant
-      var stagingFQDN = '';
+                // Use the first alias as it's the one the customer has created a CNAME over
+                let alias = matchingAliases[0];
+                debug('Using the alias %O', alias);
 
-      segments.forEach(element => {
-      // Loop through the segments rebuilding them back into a fully qualified domain name
-        stagingFQDN = stagingFQDN + element + '.';
-      });
+                // Break down the alias FQDN into its constituent parts/segments
+                let segments = alias.split('.');
+                debug('[%s] was split into these segments: %O', alias, segments);
 
-      debug('The Staging alias for %s is %s', hostname, stagingFQDN);
+                // Append '-staging' to the penultimate segment
+                segments[segments.length - 2] += '-staging';
+                debug('Edited segment %d so it is now: %O', segments.length - 2, segments);
 
-      stagingFQDN = stagingFQDN.substring(0, stagingFQDN.length - 1);
+                // Convert the array into a string storing the fully qualified Staging variant
+                let stagingFQDN = segments.join('.');
+                debug('The Staging alias for %s is %s', hostname, stagingFQDN);
 
-      debug('Calling dns.resolve(%s)', stagingFQDN);
+                debug('Calling dns.lookup(%s)', stagingFQDN);
 
-      // Resolve the Staging variant fqdn to an IP address
-      var dnsSync = require('dns-sync');
-      var stagingIPAddress = dnsSync.resolve(stagingFQDN);
+                // Set DNS lookup options
+                let options = {
+                    family: 0,
+                    verbatim: false
+                };
 
-      debug('dnsSync.resolve() returned a value of type [%s] with a value of [%O]', typeof(stagingIPAddress), stagingIPAddress);
+                // Perform lookup of staging FQDN
+                dns.lookup(stagingFQDN, options, (err, address, family) => {
 
-      if (stagingIPAddress === null){
+                    if (err) {
+                        // An error occurred
+                        debug('Error: dns.lookup() returned: %O', err);
+                        console.error('# %s failed to resolve to an IP address', stagingFQDN);
 
-        // probably not an Akamai'sed domain
-        debug('The variable [stagingIPAddress] is either null or undefined after attempting to resolve [%s]', stagingFQDN);
-        console.log('[%s] did not resolve to an IP address. [%s] is probably not served by Akamai', stagingFQDN, hostname);
-      } else {
+                    } else {
+                        debug('dns.lookup() returned - address: %j family: IPv%s', address, family);
+                        // Grab the IP Address
+                        let stagingIPAddress = address;
 
-        // Workout a string buffer length to tidy up comments' alignment
-        var entryLength = stagingIPAddress.length + hostname.length;
-        var comment = '#Akamai Staging variant of [' + alias +']';
-        const offset = 50;
-        var bufferLength = 5;
-        if(offset > entryLength + 1) {
-          bufferLength = offset - entryLength;
+                        // Define the comment we'll annotate the hosts file entry with
+                        let comment = '#Akamai Staging variant of [' + alias +']';
+
+
+                        // Workout a string buffer length (space between host file entry and comment) to tidy up comments' alignment
+                        let entryLength = stagingIPAddress.length + hostname.length;
+
+                        const offset = 50;
+                        let bufferLength = 5;
+
+                        if (offset > entryLength + 1) {
+                            bufferLength = offset - entryLength;
+                        }
+
+                        // output Staging IP entry in hosts file format
+                        console.log('%s %s %s %s', stagingIPAddress, hostname, ' '.repeat(bufferLength), comment);
+                    }
+                });
+            } else {
+              // The DNS answer didn't contain a nested domain that matched a known Akamai domain
+              console.error('# %s does not resolve to a known Akamai domain', hostname);
+            }
         }
-
-        // output Staging IP entry in hosts file format
-        console.log('%s %s %s %s', stagingIPAddress, hostname, ' '.repeat(bufferLength), comment);
-      }
-    }
-  });
-
-  debug('getStagingIPAddress() Exit');
+    });
 }
 
 try {
 
-  debug('[%s] started: %O', __filename, process.argv);
+    debug('[%s] started: %O', __filename, process.argv);
 
-  // command line options parser
-  var argv = require('yargs')
-    .help(false)
-    .argv;
+    // command line options parser
+    const argv = require('yargs')
+        .help(false)
+        .argv;
 
-  if ((process.argv.length === 2) || (argv.help)) {
-    // Show help screen
-    const help = require('./help');
-    help.helpScreen(argv.verbose);
-  } else if ((argv.platform) || (argv.info)) {
-    // Show platform information
-    const help = require('./help');
-    help.platformSpecific();
-  } else {
-    // Print a platform agnostic newline character first. This particularly helps when output is appended to a text file
-    console.log(endOfLine);
+    if ((process.argv.length === 2) || (argv.help)) {
+        // Show help screen
+        const help = require('./help');
+        help.helpScreen(argv.verbose);
 
-    // Loop through command line parameters.  Expecting 'akamai-staging.js fqdn [fqdn [fqdn] ... ]'
-    var i = 0;
-    for (i = 2; i < process.argv.length; i++) {
-      debug('Extracted "%s" from the command line', process.argv[i]);
-      // Get the Staging IP address of each fully qualified domain name specified and print to the console
-      getStagingIPAddress(process.argv[i]);
+    } else if ((argv.platform) || (argv.info)) {
+        // Show platform information
+        const help = require('./help');
+        help.platformSpecific();
+
+    } else {
+        // Print a platform agnostic newline character first. This particularly helps when output is appended to a text file
+        console.log(endOfLine);
+
+        // Loop through command line parameters.  Expecting 'akamai-staging.js fqdn [fqdn [fqdn] ... ]'
+        for (let i = 2; i < process.argv.length; i++) {
+            debug('Extracted "%s" from the command line', process.argv[i]);
+            // Get the Staging IP address of each fully qualified domain name specified and print to the console
+            getStagingIPAddress(process.argv[i]);
+        }
     }
-  }
 }
+
 catch (e) {
-  console.log('An unexpected error occurred: %s', e);
+    console.error('An unexpected error occurred: %s', e);
 }